refactor: CommunityTabs를 공용 UnderlineTabs로 추출(#712)
